+++ src/config/octane.php
@@ -1,5 +1,6 @@
 <?php
 
+use Butler\Service\Listeners\ForgetConnections;
 use Laravel\Octane\Contracts\OperationTerminated;
 use Laravel\Octane\Events\RequestHandled;
 use Laravel\Octane\Events\RequestReceived;
@@ -91,6 +92,7 @@ return [
             FlushTemporaryContainerInstances::class,
             // DisconnectFromDatabases::class,
             // CollectGarbage::class,
+            ForgetConnections::class,
         ],
 
         WorkerErrorOccurred::class => [