+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Publishable "index.php".
+- Octane listener "ForgetConnections".
 
 ### Changed
 - Require laravel v9.11.